i am getting closer. if i'm not already there. time to test!

+++ WaffleMarm/src/com/marmalade/wafflemarm/WaffleSpy.java
@@ -1,7 +1,9 @@
 package com.marmalade.wafflemarm;
+import java.io.BufferedReader;
 import java.io.BufferedWriter;
 import java.io.File;
 import java.io.FileNotFoundException;
+import java.io.FileReader;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.net.InetAddress;
@@ -23,13 +25,13 @@ public class WaffleSpy implements CommandExecutor {
 		plugin = instance;
 	}
 	static File df = WaffleMarm.plugin.getDataFolder();
-	
+
 	@EventHandler(priority = EventPriority.LOWEST)
 	public void AsyncPlayerPreLoginEvent(AsyncPlayerPreLoginEvent e){
 		InetAddress ip = e.getAddress();
 		String iP = ip.getHostName();
 		String playerName = e.getName();
-		OfflinePlayer offlinePlayer = plugin.getServer().getOfflinePlayer(null);
+		OfflinePlayer offlinePlayer = plugin.getServer().getOfflinePlayer(playerName);
 		long fp = offlinePlayer.getFirstPlayed();
 		long lp = offlinePlayer.getLastPlayed();
 		File waffleInfo = new File("C:\\WaffleMarm\'players.yml");
@@ -38,17 +40,19 @@ public class WaffleSpy implements CommandExecutor {
 		} catch (IOException err) {
 			err.printStackTrace();
 		}
-	    try {
-	    	BufferedWriter writer = new BufferedWriter(new FileWriter(waffleInfo));
-	    	writer.write("Username: " + playerName + ", " + "IP: " + iP);
-	    	writer.newLine();
-	    	writer.write("First Login: " + fp + ", " + "Last Login: " + lp);
-	    	writer.flush();
-	    	writer.close();
-		} catch (FileNotFoundException err) {
-			System.out.println("File Not Found. Good Luck!");
-		} catch (IOException err) {
-			err.printStackTrace();
+		if(waffleInfo.exists()){
+			try {
+				BufferedWriter writer = new BufferedWriter(new FileWriter(waffleInfo));
+				writer.write("Username: " + playerName + ", " + "IP: " + iP);
+				writer.newLine();
+				writer.write("First Login: " + fp + ", " + "Last Login: " + lp);
+				writer.flush();
+				writer.close();
+			} catch (FileNotFoundException err) {
+				System.out.println("File Not Found. Good Luck!");
+			} catch (IOException err) {
+				err.printStackTrace();
+			}
 		}
 	}
 
@@ -56,10 +60,34 @@ public class WaffleSpy implements CommandExecutor {
 		Player player = (Player) sender;
 		String commandName = command.getName().toLowerCase();
 		String[] trimmedArgs = args;
+		OfflinePlayer offlinePlayer = plugin.getServer().getOfflinePlayer(trimmedArgs[0]);
+		long fp = offlinePlayer.getFirstPlayed();
+		File waffleInfo = new File("C:\\WaffleMarm\'players.yml");
+		try {
+			if(waffleInfo.exists()){
+				BufferedReader reader = new BufferedReader(new FileReader(waffleInfo));
+				String str = reader.readLine();
+				while(str != null){
+					sender.sendMessage(str);
+					reader.close();
+				}
+			}
+		} catch (FileNotFoundException e) {
+			if(fp == 0){
+				System.out.println("Player Does Not Exist!");
+			}
+			else{
+				System.out.println("File Not Found!");
+				e.printStackTrace();
+			}
+		} catch (IOException e) {
+			e.printStackTrace();
+		}
 		if (commandName.equalsIgnoreCase("seen " + trimmedArgs[0])) {
 			if(player.isOp()){
 				if (trimmedArgs.length == 1) {
-					sender.sendMessage(ChatColor.GOLD + "Player information for " + ChatColor.GREEN + trimmedArgs[0]);		
+					sender.sendMessage(ChatColor.GOLD + "Player information for " + ChatColor.GREEN + trimmedArgs[0]);
+
 				}
 			}
 		}
